refactor: cleanup of in-progress work
