+++ internal/providers/url_template_datatype_test.go
@@ -1,0 +1,29 @@
+// Copyright 2026 Michael Davis
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package providers
+
+import (
+	"testing"
+
+	"github.com/Michad/tilegroxy/pkg"
+	"github.com/stretchr/testify/assert"
+)
+
+// URLTemplate embeds Proxy, so it inherits DataType() without its own method. It needs no
+// external process, so it's kept out of url_template_test.go's !unit build tag and runs as
+// part of the regular unit suite.
+func Test_DataType_URLTemplate_InheritsFromProxy(t *testing.T) {
+	assert.Equal(t, pkg.DataTypeUnknown, URLTemplate{}.DataType())
+}
